Clean up project.
