--- PizzaHat/config.yml
@@ -1,2 +1,0 @@
-owners: 710247495334232164
-guild_id: 764049436275114004
